Cleaning and logging

+++ UnityProject/Assets/OpenMined/Syft/Tensor/FloatTensor.Autograd.cs
@@ -35,7 +35,6 @@ namespace OpenMined.Syft.Tensor
             return true;
         }
 
-
         // hook autograd two parents - one scalar
         public void HookAutograd(ref FloatTensor result, float x, string creation_op)
         {
@@ -44,22 +43,21 @@ namespace OpenMined.Syft.Tensor
                 FloatTensor new_child =
                     new FloatTensor(_controller: controller, _shape: new int[] {1}, _data: new float[] {x});
 
-                result.InitAutograd();
-                result.creators.Add(this);
-                result.creators.Add(new_child);
+                result.InitAutograd ();
+                result.creators.Add (this);
+                result.creators.Add (new_child);
                 result.creation_op = creation_op;
 
-                children.Add(result.Id, 0);
+                children.Add (result.Id, 0);
 //				new_child.children.Add (result.Id, 0);
 			}
-
 		}
 
 		// hook autograd two parents
-		public void HookAutograd(ref FloatTensor result, ref FloatTensor x, string creation_op) {
-
-			if (autograd) {
-
+		public void HookAutograd (ref FloatTensor result, ref FloatTensor x, string creation_op)
+        {
+			if (autograd)
+            {
 				result.InitAutograd ();
 				result.creators.Add (this);
 				result.creators.Add (x);
@@ -67,16 +65,14 @@ namespace OpenMined.Syft.Tensor
 
 				children.Add (result.Id, 0);
 				x.children.Add (result.Id, 0);
-
 			}
-
 		}
 
 		// hook autograd single parent
-		public void HookAutograd(ref FloatTensor result, string creation_op) {
-
-			if (autograd) {
-
+		public void HookAutograd (ref FloatTensor result, string creation_op)
+        {
+			if (autograd)
+            {
 				result.InitAutograd ();
 				result.creators.Add (this);
 				result.creation_op = creation_op;
@@ -85,14 +81,13 @@ namespace OpenMined.Syft.Tensor
 			}
 		}
 
-	    public void Backward(FloatTensor grad = null, FloatTensor grad_origin = null)
+	    public void Backward (FloatTensor grad = null, FloatTensor grad_origin = null)
 	    {
-
 		    if (autograd)
 		    {
 			    if (grad == null)
 			    {
-				    grad = this.controller.createOnesTensorLike(this);
+				    grad = this.controller.createOnesTensorLike (this);
 				    grad.Autograd = false;
 			    }
 
@@ -114,7 +109,7 @@ namespace OpenMined.Syft.Tensor
 			    }
 			    else
 			    {
-				    this.Grad.Add(grad, true);
+				    this.Grad.Add (grad, true);
 			    }
 
 			    // grads must not have grads of their own
@@ -130,64 +125,47 @@ namespace OpenMined.Syft.Tensor
 			    {
 				    if (creation_op == "add_elem")
 				    {
-
 					    creators[0].Backward(grad.Copy(), this);
 					    creators[1].Backward(grad.Copy(), this);
-
 				    }
 				    else if (creation_op == "mul_elem")
 				    {
-
 					    creators[0].Backward(grad.Mul(creators[1]), this);
 					    creators[1].Backward(grad.Mul(creators[0]), this);
-
 				    }
 				    else if (creation_op == "div_elem")
 				    {
-
 					    creators[0].Backward(grad.Div(creators[1]), this);
 					    creators[1].Backward(grad.Div(creators[0]), this);
-
 				    }
 				    else if (creation_op == "sub_elem")
 				    {
-
 					    creators[0].Backward(grad.Copy(), this);
 					    creators[1].Backward(grad.Neg(), this);
-
 				    }
 				    else if (creation_op == "mm")
 				    {
-
 					    creators[0].Backward(grad.MM(creators[1].Transpose()).Transpose(), this);
 					    creators[1].Backward(grad.Transpose().MM(creators[0]).Transpose(), this);
-
 					    /*creators [0].Backward (creators [1].MM (grad.Transpose ()), this);
 					    creators [1].Backward (creators [0].Transpose ().MM (grad), this);*/
-
 				    }
 				    else if (creation_op == "sigmoid")
 				    {
-
 					    FloatTensor c = this.Copy();
 					    c.autograd = false;
-					    creators[0].Backward(c.Neg().Add((float) 1).Mul(this).Mul(grad), this);
-
+					    creators[0].Backward(c.Neg().Add(1.0f).Mul(this).Mul(grad), this);
 				    }
 				    else if (creation_op == "pow_scalar")
 				    {
-
 					    FloatTensor self_nograd = creators[0].Copy();
 					    self_nograd.autograd = false;
 					    creators[0].Backward(self_nograd.Mul(grad).Mul(creators[1].Data[0]), this);
-
 				    }
-
+                    
 //					if (!keepgrads) {
 //						ctrl.RemoveTensor (grad.id);
 //					}
-
-
 			    }
 		    }
 	    }
